feat: async notifier 移行 / weather
