+++ app/melos_template/.env.local
@@ -1,0 +1,1 @@
+OPEN_WEATHER_API_KEY=""
