+++ app/melos_template/lib/feature/home/home_page.dart
@@ -16,7 +16,7 @@ class HomePage extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final vm = ref.watch(homeViewmodelProvider.notifier);
-    final state = ref.watch(homeViewmodelProvider);
+
     final notificationService = ref.read(notificationServiceProvider);
 
     return AppBarFrame(
@@ -42,86 +42,82 @@ class HomePage extends ConsumerWidget {
         onTap: () {
           context.hideKeyboard();
         },
-        child: !state.isInternetConnected
-            ? AppRetryNetworkConnectWidget(
-                onRetry: vm.checkInternetConnection,
-              )
-            : Center(
-                child: Padding(
-                  padding: const EdgeInsets.all(16),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      const Text(
-                        'Welcome to the App!',
-                        style: TextStyle(
-                          fontSize: 24,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                      ElevatedButton.icon(
-                        onPressed: () {
-                          notificationService.showNotification(
-                            title: 'Hello!',
-                            body: 'This is a notification from the app!',
-                          );
-                          const E2ESamplePageData().push<void>(context);
-                        },
-                        icon: const Icon(
-                          Icons.sunny,
-                          color: Colors.red,
-                          size: 24,
-                        ),
-                        label: const Text(
-                          'E2E Sample',
-                          style: TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        style: ElevatedButton.styleFrom(
-                          padding: const EdgeInsets.symmetric(
-                            vertical: 14,
-                            horizontal: 20,
-                          ),
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(12),
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                      ElevatedButton.icon(
-                        onPressed: () {
-                          const WeatureRouteData().push<void>(context);
-                        },
-                        icon: const Icon(
-                          Icons.sunny,
-                          color: Colors.red,
-                          size: 24,
-                        ),
-                        label: const Text(
-                          'Get to weather info (REST API)',
-                          style: TextStyle(
-                            fontSize: 18,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        style: ElevatedButton.styleFrom(
-                          padding: const EdgeInsets.symmetric(
-                            vertical: 14,
-                            horizontal: 20,
-                          ),
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(12),
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 16),
-                    ],
+        child: Center(
+          child: Padding(
+            padding: const EdgeInsets.all(16),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const Text(
+                  'Welcome to the App!',
+                  style: TextStyle(
+                    fontSize: 24,
+                    fontWeight: FontWeight.bold,
                   ),
                 ),
-              ),
+                const SizedBox(height: 16),
+                ElevatedButton.icon(
+                  onPressed: () {
+                    notificationService.showNotification(
+                      title: 'Hello!',
+                      body: 'This is a notification from the app!',
+                    );
+                    const E2ESamplePageData().push<void>(context);
+                  },
+                  icon: const Icon(
+                    Icons.sunny,
+                    color: Colors.red,
+                    size: 24,
+                  ),
+                  label: const Text(
+                    'E2E Sample',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  style: ElevatedButton.styleFrom(
+                    padding: const EdgeInsets.symmetric(
+                      vertical: 14,
+                      horizontal: 20,
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                ),
+                const SizedBox(height: 16),
+                ElevatedButton.icon(
+                  onPressed: () {
+                    const WeatherRouteData().push<void>(context);
+                  },
+                  icon: const Icon(
+                    Icons.sunny,
+                    color: Colors.red,
+                    size: 24,
+                  ),
+                  label: const Text(
+                    'Get to weather info (REST API)',
+                    style: TextStyle(
+                      fontSize: 18,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  style: ElevatedButton.styleFrom(
+                    padding: const EdgeInsets.symmetric(
+                      vertical: 14,
+                      horizontal: 20,
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                ),
+                const SizedBox(height: 16),
+              ],
+            ),
+          ),
+        ),
       ),
     );
   }
