+++ app/melos_template/pubspec.yaml
@@ -28,12 +28,10 @@ dependencies:
     path: ../../packages/test_util
   core_network:
     path: ../../packages/network
-
   introduction_screen: ^3.1.14
   timeago: ^3.7.0
   mockito: ^5.4.4
   dio: ^5.7.0
-
   firebase_analytics: ^11.3.3
   firebase_core: ^3.6.0
   firebase_messaging: ^15.0.0
@@ -57,6 +55,7 @@ dependencies:
   connectivity_plus: ^6.1.0
   patrol: ^3.13.0
   dotenv: ^4.2.0
+  flutter_dotenv: ^5.2.1
 
 dev_dependencies:
   cider: ^0.2.8
@@ -70,6 +69,7 @@ dev_dependencies:
 flutter:
   uses-material-design: true
   assets:
+    - .env
     - assets/images/icons/
     - assets/images/introduction/
   fonts:
